same parameter name for USER_KEY
It was named USER_KEY in the argument list but used as USER_KEY_PATH.
The _PATH suffix probably originates from the same parameter for the
sign subcommand, which doesn't make sense here though as it only handles
the revocation.

A more elaborate fix would be to name everything similarly, but that has
a higher chance of introducing breakage.

Signed-off-by: benaryorg <binary@benary.org>